src-python/requirements.txt
###### Requirements without Version Specifiers ######
torch
numpy
jax
flax
wheel
transformers
progressbar2